Change SpdyFrameBuilder::length() to return size_t instead of int to avoid a bunch of static_casts in SpdyFramer.
Review URL: http://codereview.chromium.org/10162008

git-svn-id: svn://svn.chromium.org/chrome/trunk/src@133321 0039d316-1c4b-4281-b951-d872f2087c98

+++ net/spdy/spdy_frame_builder.h
@@ -39,7 +39,7 @@ class NET_EXPORT_PRIVATE SpdyFrameBuilder {
   SpdyFrameBuilder(SpdyStreamId stream_id, SpdyDataFlags flags,  size_t size);
 
   // Returns the size of the SpdyFrameBuilder's data.
-  int length() const { return length_; }
+  size_t length() const { return length_; }
 
   // Takes the buffer from the SpdyFrameBuilder.
   SpdyFrame* take() {
